requirements.txt
streamlit==1.38.0
python-frontmatter==1.0.1
